+++ javaScript/projects/knights-travails/src/containers/binary-search-tree/node.js
@@ -1,7 +1,0 @@
-export default class Node {
-  constructor(data) {
-    this.leftNode = null;
-    this.rightNode = null;
-    this.data = data !== null ? data : null;
-  }
-}