+++ javaScript/projects/knights-travails/src/containers/node.js
@@ -1,0 +1,8 @@
+export default class Node {
+  constructor(value, depth, possibleMoves, previous) {
+    this.value = value;
+    this.depth = depth;
+    this.possibleMoves = possibleMoves;
+    this.previous = previous;
+  }
+}